Fix multiple-definition and type-mismatch linker errors
- src/hazard/hazrd4.h: remove 'nrow' from HAZRD4 definitions block;
  nrow is already defined in hazard.c via hazard.h (EXTERN int prnt,nrow)
  so having it here too caused a multiple definition error on strict linkers

- src/hazard/hazard_l.c: change 'int yylnctr,yysynerr' to
  'extern int yylnctr; int yysynerr;' — yylnctr is already defined in
  hazard.c via structures.h (EXTERN int yylnctr with HAZARD defined);
  yysynerr remains a definition here as it is unique to the lexer

- src/hazard/setstat.c + setstat.h: unify function signature to
  int* for all three pointer params; the implementation had
  (long*,long*,int*) while the header had (int*,int*,long*) — both wrong.
  All callers pass int* (H->phase1 is int, iearly/iconst/ilate are int*,
  iexc/icxc/ilxc are int), so int* throughout is correct.
  The mismatched signature was an error on Windows/MinGW, warning on Linux.

+++ src/hazard/setstat.c
@@ -4,8 +4,8 @@
 #include <isblanck.h>
 /****************************************************************/
 /* SETSTAT called from SETCONC                                  */
-void setstat(int i,int idx,long *phasevar,long *itabl,
-		    int *ipxc){
+void setstat(int i,int idx,int *phasevar,int *itabl,
+                    int *ipxc){
   if((memcmp(faze[i],"EARLY   ",8)==0 && C->phase[1]!=1) ||
      (memcmp(faze[i],"CONSTANT",8)==0 && C->phase[2]!=1) ||
      (memcmp(faze[i],"LATE    ",8)==0 && C->phase[3]!=1))
